discord-bot/pyproject.toml
[project]
name = "discord-bot"
version = "0.1.0"
description = "Discord bot extracted from agent-coordinator"
requires-python = ">=3.10"
dependencies = [
    "discord.py>=2.4.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
]

[project.scripts]
discord-bot = "coordinator.discord.bot:run_bot"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["coordinator"]

[tool.uv]
dev-dependencies = []

